rename generated packages
