+++ index/segment/mem/postingsgen/concurrent_map_test.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package postingsmap
+package postingsgen
 
 import (
 	"regexp"
