+++ index/segment/mem/postingsgen/generated_map.go
@@ -22,7 +22,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package postingsmap
+package postingsgen
 
 import "github.com/m3db/m3ninx/postings"
